requirements.txt
repsys-framework==0.3.3
scikit-learn
pandas
scipy
numpy
